Harden testing
